[Table Vis] add an experimental table visualization in vis builder
To convert the table visualization into React & OUI DataGrid component,
in this PR, we did two main things:
* add an experimental table visualization
* enable it in vis builder

The experimental table vis has all the features from current table. It
does couple of things:
* restore table vis in react using OUI datagrid component
* datagrid component does not support splitted grids. For future transfer
to OUI Datagrid, we create a tableGroup in visData for splitted grids.
* restore basic pagenation, sort and format.
* implement datagrid columns
* display column title correctly
* deangular and re-use formatted column
* convert formatted column to data grid column
* restore filter in and filter out value functions
* make table vis column resizable
* format table cell to show Date and percent
* restore showTotal feature: it allows table vis to show total,
avg, min, max and count statics on count
* restore export csv feature to table vis
* split table in rows and columns

Issue Resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2212
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2213
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2379

[Table Visualization][IMPROVE] remove repeated column from split tables (#2583)

Currently, when we split table by columns, the split column is shown
both in the table title and as a separate column. This is not needed.
In this PR, we remove the repeated column in split tables in col.

Partically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2579#issuecomment-1281099947

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization][BUG] remove local column width state (#2582)

* remove local col width state to allow split tables to fetch
updated col width state
* fix type errors in usePagination

Partially resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2579#issuecomment-1281078065

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization][BUG] partical rows shows metrics for all columns (#2648)

Currently, when we enable Show partial rows in the Options panel, we see metrics been
added to every column even though Show metrics for every bucket/level is not enabled.

Metrics are added and returned when we enable the partial rows. This PR fixed the bug
by slice the returned data to allow only the last set of metrices.

Partially resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2579#issuecomment-1287435381

Signed-off-by: Anan Zhuang <ananzh@amazon.com>