[Table Visualization] remove legacy table and add table visualization (#2279)
To convert the table visualization into React & OUI DataGrid component,
in this PR, we did two main things:
* clean out legacy angular code
* restore table vis in react
* Datagrid component does not support splitted grids. For future transfer
to OUI Datagrid, we create a tableGroup in visData for splitted grids.

issue resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2212
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2213a

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

rename visTable back to opensearch_dashboards_table and add stronger type

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] add a plain datagrid component (#2390)

implement a plain OuiDataGrid component use the basic
pagenation, sort and format.

Partially resolve:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] restore datagrid columns (#2411)

* restore datagrid columns
* display column title correctly
* deangular and re-use formatted column
* convert formatted column to data grid column
* restore filter in and filter out value functions

Partially resolve:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] restore pagination to table vis (#2461)

* add pagination

patically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] make table vis column resizable (#2464)

* add resizable state to column

Partially resolve:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] make table vis column sortable (#2502)

* add sort state (asc | desc)  to column
* fix pagination issue

Partially resolve:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2305

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] format table cell and restore showTotal feature (#2562)

* format table cell to show Date and percent
* restore showTotal feature: it allows table vis to show total,
avg, min, max and count statics on count
* fix some type errors

Partically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2379

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization]restore export csv feature to table vis (#2568)

* add addtional action in toolbar to allow export data to csv.
there are two types of csv, raw and formatted. raw is the original
data and formatted is to show formatted Date and percentage data
when needed.
* when table is not saved, export csv file will be named as
unsaved-raw.csv if choose raw. when table is saved with a filename,
it will be saved as [filename]-[raw/formatted].csv

Partically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2379

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization] split table in rows and columns (#2578)

This PR implement a group component TableVisGroupComponent utilizing
TableVisComponent as sub component. It also adds a title to TableVisComponent.

Partically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2379

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization][IMPROVE] remove repeated column from split tables (#2583)

Currently, when we split table by columns, the split column is shown
both in the table title and as a separate column. This is not needed.
In this PR, we remove the repeated column in split tables in col.

Partically resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2579#issuecomment-1281099947

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization][BUG] remove local column width state (#2582)

* remove local col width state to allow split tables to fetch
updated col width state
* fix type errors in usePagination

Partially resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2579#issuecomment-1281078065

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

Signed-off-by: Anan Zhuang <ananzh@amazon.com>

[Table Visualization][BUG] partical rows shows metrics for all columns (#2648)

Currently, when we enable Show partial rows in the Options panel, we see metrics been
added to every column even though Show metrics for every bucket/level is not enabled.

Metrics are added and returned when we enable the partial rows. This PR fixed the bug
by slice the returned data to allow only the last set of metrices.

Partially resolved:
https://github.com/opensearch-project/OpenSearch-Dashboards/issues/2579#issuecomment-1287435381

Signed-off-by: Anan Zhuang <ananzh@amazon.com>